completed new auth sheet
